Bunch of work towards rasterizing random shapes and debug

+++ Converter/FileStructures/Type1/Data.cs
@@ -2,19 +2,19 @@
 {
   public class TYPE1_Point2D
   {
-    public float X;
-    public float Y;
+    public double X;
+    public double Y;
     public bool PartOfPath;
     public TYPE1_Point2D() { }
 
-    public TYPE1_Point2D(float x, float y)
+    public TYPE1_Point2D(double x, double y)
     {
       X = x;
       Y = y;
       PartOfPath = false;
     }
 
-    public TYPE1_Point2D(int x, int y, bool partOfPath)
+    public TYPE1_Point2D(double x, double y, bool partOfPath)
     {
       X = x;
       Y = y;
